add final check for node join cluster
